requirements.txt
fastapi
uvicorn
streamlit
PyJWT
requests
typing-extensions
langgraph
langgraph-checkpoint
langgraph-prebuilt
langchain
langchain-community
langchain-ollama
langchain-cli
pyperclip
fastapi
sentence_transformers
scipy
networkx
sqlglot
langchain-mcp-adapters
sentence_transformers
langgraph-prebuilt==0.5.2
langchain-core
mcp==1.11.0
langchain_mcp_adapters
langchain_community
langchain_openai